+++ CHANGELOG.md
@@ -33,6 +33,8 @@ Released on ??
 
 > 🦥 The lazy update
 
+- **Default ssh config path**:
+  - SSH configuration path is now `~/.ssh/config` by default
 - **Bugfix**:
   - Fixed [Issue 126](https://github.com/veeso/termscp/issues/126)
   - Fixed [Issue 141](https://github.com/veeso/termscp/issues/141)